feat: 프로젝트 세부 모달
