+++ apps/ceos/src/components/project/ProjectCardContainer/index.tsx
@@ -1,5 +1,24 @@
 import { ProjectCard, ProjectCardProps } from '@ceos-fe/ui';
+import DetailModal from '@ceos/pages/project/DetailModal';
+import { useState } from 'react';
 
 export const ProjectCardContainer = (props: ProjectCardProps) => {
-  return <ProjectCard {...props} />;
+  const [modalNumber, setModalNumber] = useState(-1);
+  const setClose = () => {
+    setModalNumber(-1);
+  };
+  const openModal = () => {
+    setModalNumber(props.id);
+    console.log(props.id);
+  };
+  return (
+    <>
+      <div onClick={openModal}>
+        <ProjectCard {...props} />
+      </div>
+      {modalNumber !== -1 && (
+        <DetailModal id={modalNumber} setClose={setClose} />
+      )}
+    </>
+  );
 };
